Slippage function (#319)
* add new slippage variables and setters

* add new params to test function calls

* build slippage function in pricer

* add local test function to calculate slippage

* update netDhvExposure state

* update liquidity pool tests for slippage

* add natspec to slippage func

* remove console logs

* merge rysk-beyond into branch

* replace delta bands with a dynamic array system

* update local slippage function and tests

* convert slippage into exponential function

* remove unused vars from struct

* fix exposureCoefficient calculation and modify tests

* fix: fix merge conflict issues

* chore: fix tests and reduce size

* chore: add slippage gradient tests to prove works

* fix: fix bugs and continue fixing tests

* Feat/test slippage function (#325)

* chore: add slippage gradient tests to prove works

* fix: fix bugs and continue fixing tests

* feat: finish fixing tests

* feat: fix test bug

* Feat/slippage changes (#332)

* chore: add slippage gradient tests to prove works

* fix: fix bugs and continue fixing tests

* feat: finish fixing tests

* feat: fix test bug

* feat: new slippage multiplier funciton

* feat: create foundry ffi function and python function

* fix: final bugs

* feat: add quote checks

* fix: tests

* feat: add minimum amount

* feat: add foundry fuzz testing on slippage function

* Spread function (#334)

* add collateral lending rate var

* add max function to optionsCompute

* build spread function for short options

* fix errors on compile:

* linting

* build local spread func

* update spread function and pricer deploy params

* set up spread test suite

* fix test suite decimals issue and apply spread to total contracts quote not individual

* undo commented out tests

* swap long and short delta borrow rates

* remove unused max function

* fix onlyGuardian funcs to onlyGovernon

* fix comment

* chore: repair test


* refactor: move issuance and netdhv logic to catalogue (#336)

* refactor: move issuance and netdhv logic to catalogue

* chore: fix all tests

* feat: add test for otc and exchange trade on same book

* feat: add min and max for trade size and netdhvexposure

* chore: add comment to updateDhvExposure

* lint

* lint

* chore: get better cov

* chore: fix slippage bug

* feat: update docs

* Spread function (#339)

* add collateral lending rate var

* add max function to optionsCompute

* build spread function for short options

* fix errors on compile:

* linting

* build local spread func

* update spread function and pricer deploy params

* set up spread test suite

* fix test suite decimals issue and apply spread to total contracts quote not individual

* undo commented out tests

* swap long and short delta borrow rates

* remove unused max function

* fix onlyGuardian funcs to onlyGovernon

* fix comment

* chore: repair test

* add spread foundry test contract

* add more test permutations to foundry fuzz

* fuzz state vars

* add state var fuzzing

* add more permutations to slippage foundry fuzz

+++ packages/contracts/contracts/libraries/OptionsCompute.sol
@@ -196,13 +196,13 @@ library OptionsCompute {
 		);
 	}
 
-	function min(uint v1, uint v2) internal pure returns (uint) {
+	function min(uint256 v1, uint256 v2) internal pure returns (uint256) {
 		return v1 > v2 ? v2 : v1;
 	}
-	
+
 	function toInt256(uint256 value) internal pure returns (int256) {
-        // Note: Unsafe cast below is okay because `type(int256).max` is guaranteed to be positive
-        require(value <= uint256(type(int256).max), "SafeCast: value doesn't fit in an int256");
-        return int256(value);
-    }
+		// Note: Unsafe cast below is okay because `type(int256).max` is guaranteed to be positive
+		require(value <= uint256(type(int256).max), "SafeCast: value doesn't fit in an int256");
+		return int256(value);
+	}
 }
